add repositories unit for all model

+++ Task5_EF/Repositories/FlowerWarehouseRepository.cs
@@ -1,13 +1,12 @@
 ﻿using System.Data.Entity;
 using System.Collections.Generic;
 using System.Linq;
-using Task5_EF.Interfaces;
 using Task5_EF.IRepository;
 using Task5_EF.Models;
 
 namespace Task5_EF.Repositories
 {
-    public class FlowerWarehouseRepository :IRepository<FlowerWarehouseModel>
+    public class FlowerWarehouseRepository : IRepository<FlowerWarehouseModel>
     {
         private SupplyContext supplyDbContext;
 
@@ -18,20 +17,29 @@ namespace Task5_EF.Repositories
 
         public void Create(FlowerWarehouseModel item)
         {
-            supplyDbContext.FlowerWarehouses.Add(item);
+            if (item != null)
+            {
+                supplyDbContext.FlowerWarehouses.Add(item);
+            }
         }
 
         public void Delete(FlowerWarehouseModel item)
         {
-            var deleteItem = GetById(item.Id);
-            supplyDbContext.FlowerWarehouses.Remove(deleteItem);
+            if (item != null)
+            {
+                var deleteItem = GetById(item.Id);
+                if (deleteItem != null)
+                {
+                    supplyDbContext.FlowerWarehouses.Remove(deleteItem);
+                }
+            }
         }
 
         public List<FlowerWarehouseModel> GetAll()
         {
             return supplyDbContext.FlowerWarehouses
-                .Include(x=>x.FlowerModel)
-                .Include(x=>x.WarehouseModel)
+                .Include(x => x.FlowerModel)
+                .Include(x => x.WarehouseModel)
                 .ToList();
         }
 
@@ -43,8 +51,14 @@ namespace Task5_EF.Repositories
 
         public void Update(FlowerWarehouseModel item)
         {
-            var updateItem = GetById(item.Id);
-            updateItem.Amount = item.Amount;
+            if (item != null)
+            {
+                var updateItem = GetById(item.Id);
+                if (updateItem != null)
+                {
+                    updateItem.Amount = item.Amount;
+                }
+            }
         }
     }
 }
